% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_index_to_vector}
\alias{add_index_to_vector}
\title{Return list with vector and indices}
\usage{
add_index_to_vector(increment_pos, vec)
}
\arguments{
\item{increment_pos}{Positions where index increment by one, the range of the returned vectors is min:max of this vector}

\item{vec}{Vector to add indices too}
}
\value{
data.table with read mappings
}
\description{
Return list with vector and indices
}
