% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_mapping}
\alias{add_mapping}
\title{Add read mappings to metainfo}
\usage{
add_mapping(metainfo, read_mapping)
}
\arguments{
\item{metainfo}{data.table of metainfo loaded with read_metainfo}

\item{read_mapping}{read mappings}
}
\value{
Nothing, signal mappings will be added to the metainfo object in memory
}
\description{
Add necessary read mapping information to metainfo
}
