% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_signal}
\alias{add_signal}
\title{Add signal mapping to metainfo}
\usage{
add_signal(metainfo, hdf5_obj, batch)
}
\arguments{
\item{metainfo}{data.table of metainfo loaded with read_metainfo}

\item{hdf5_obj}{Open hdf5 object}

\item{batch}{Batch to load signal mappings from}
}
\value{
Nothing, signal mappings will be added to the metainfo object in memory
}
\description{
Loads the signal mappings associated with a batch
}
