% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_detection.R
\name{calculate_bit_score}
\alias{calculate_bit_score}
\title{Calculate bit score and extract motifs}
\usage{
calculate_bit_score(
  sequences,
  nucleotides = c("G", "C", "A", "T"),
  min_entropy = 0.4
)
}
\arguments{
\item{sequences}{list of sequences of the same length}
}
\value{
list(entropy_dt, motif_dt, motif_vec)
}
\description{
The entropy at each postion in supplied sequences. The entropy
calculated based on
https://en.wikipedia.org/wiki/Sequence_logo#Logo_creation.
}
