% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mappings.R
\name{chunk_mapping}
\alias{chunk_mapping}
\title{Chunk mapped reads into regions}
\usage{
chunk_mapping(read_mapping, chunk_size = 1e+05)
}
\arguments{
\item{read_mapping}{read_mapping loaded with 'load_mapping'}

\item{chunk_size}{Size of chunks to divide contigs into}
}
\value{
Nothing, chunks get added to dt in memory
}
\description{
Divides mapped read into chunks based on position on refernce mapped too.
}
