% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_detection.R
\name{embed_and_cluster}
\alias{embed_and_cluster}
\title{Embed joined chunks}
\usage{
embed_and_cluster(
  chunks,
  ref_path,
  out,
  event_pval_limit = 1e-30,
  event_sequence_frame = 5,
  umap_args = list(),
  hdbscan_args = list()
)
}
\arguments{
\item{chunks}{preprocessed chunks generated with preprocess_all_chunks}

\item{ref_path}{path to the reference passed to megalodon}

\item{out}{character string with path to output results}

\item{event_pval_limit}{Numeric, max p-value for events}

\item{event_sequence_frame}{integer, number of nucleotide to include to each side}

\item{umap_args}{list of arguments passed directly to umap (See ?umap::umap)
https://cran.r-project.org/web/packages/umap/vignettes/umap.html#configuration-objects}

\item{hdbscan_Args}{list of arguments passed directly to hdbscan (See ?dbscan::hdbscan)}
}
\value{
embedded and clustered positions
}
\description{
Wrapper function for umbedding and clustering. Take the data.table containing
the output the preprocess_chunk function. Precessed chunks can be loaded using
load_processed_chunks
}
