% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_detection.R
\name{find_motifs}
\alias{find_motifs}
\title{Visualise motif profiles}
\usage{
find_motifs(
  path_chunk_stats,
  path_ref,
  out,
  umap_args = list(n_components = 2),
  hdbscan_args = list(minPts = 30),
  p_value_threshold = 1e-20,
  event_sequence_frame = 8,
  align_event_sequences = TRUE,
  iterations = 5,
  iteration_approach = "remove_noise"
)
}
\arguments{
\item{path_chunk_stats}{path to preprocess chunks}

\item{path_ref}{path to reference}

\item{out}{output folder}

\item{n_extra_positions}{number of adjacent positions to include in plot}

\item{motifs_evaluated}{evaluation type (all, clustered, or vector of motifs)}
}
\value{
plots
}
\description{
Visualise the identified clusters
}
