% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_hdf_batches}
\alias{get_hdf_batches}
\title{See which batches are present in signal_mappings.hdf}
\usage{
get_hdf_batches(hdf5)
}
\arguments{
\item{hdf5}{Open hdf5 object}
}
\value{
vector with batch names
}
\description{
See which batches are present in signal_mappings.hdf
}
