% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mappings.R
\name{get_overextending_reads}
\alias{get_overextending_reads}
\title{Get read mappings that overextent their chunk}
\usage{
get_overextending_reads(read_mapping, chunk_size)
}
\arguments{
\item{read_mapping}{Read mapping dt}

\item{chunk_size}{Size of chunks}
}
\value{
read mapping data.table
}
\description{
Gets the read that overextent their assigned chunk and return dt with new chunk assignments
}
