% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_difference.R
\name{get_reference_context}
\alias{get_reference_context}
\title{Load signal mappings of one batch}
\usage{
get_reference_context(signal_dt, chunk_size)
}
\arguments{
\item{chunk_size}{integer, size of chunk}

\item{signal}{data.table of metainfo loaded with load_metainfo}
}
\value{
data.table
}
\description{
Loads the signal mappings associated with a batch
}
