% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_detection.R
\name{load_processed_chunks}
\alias{load_processed_chunks}
\title{Load preprocessed chunks}
\usage{
load_processed_chunks(preprocess_out)
}
\arguments{
\item{preprocess_out}{path specified by 'out' in the preprocess function}
}
\value{
joined chunks
}
\description{
Loads the output of the preprocess function
}
