% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{load_read_mapping}
\alias{load_read_mapping}
\title{Load sorted read mappings}
\usage{
load_read_mapping(path_mapping)
}
\arguments{
\item{path_mapping}{Preprocessed signal mappings}
}
\value{
read mapping data.table
}
\description{
Load read mapping required for correct assignment of signal mapping reference.
}
