% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{load_signal_mapping_metainfo}
\alias{load_signal_mapping_metainfo}
\title{Read metainfo from signal_mappings.hdf}
\usage{
load_signal_mapping_metainfo(hdf5, batch)
}
\arguments{
\item{hdf5}{Open hdf5 object}

\item{batch}{Batch to read from}
}
\value{
data.table with metainformation
}
\description{
This function load metainformation of each read in a batch in a singal mapping hdf5 file.
}
