% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{load_signal_mappings_batch}
\alias{load_signal_mappings_batch}
\title{Load signal mappings of one batch}
\usage{
load_signal_mappings_batch(metainfo, hdf5, batch)
}
\arguments{
\item{metainfo}{data.table of metainfo loaded with metainfo_load_batch}

\item{hdf5}{Open hdf5 object}
}
\value{
Nothing, signal mappings will be added to the metainfo object in memory
}
\description{
Loads the signal mappings associated with a batch
}
