% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{metainfo_load_all_batches}
\alias{metainfo_load_all_batches}
\title{Read metainfo from signal_mappings.hdf}
\usage{
metainfo_load_all_batches(hdf5)
}
\arguments{
\item{hdf5}{Open hdf5 object}
}
\value{
data.table with metainformation of all batches
}
\description{
This function load all  metainformation in a singal mapping hdf5 file.
}
