% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanomotif.R
\name{nanomotif}
\alias{nanomotif}
\title{Nanomotif}
\usage{
nanomotif(
  nat_mapping,
  pcr_mapping,
  nat_signal,
  pcr_signal,
  out,
  reference_path,
  chunk_size = 1e+05,
  max_cov = 200,
  minPval = 1e-15,
  hdbscan_args = list(minPts = 10),
  umap_args = list(n_components = 2, min_dist = 0.05, n_neighbors = 30),
  align_event_sequences = TRUE,
  iterations = 1,
  min_entropy = 1,
  chunks_per_contig = 0
)
}
\arguments{
\item{nat_mapping}{Path to NAT sorted bam file}

\item{pcr_mapping}{Path to PCR sorted bam file}

\item{nat_signal}{Path to NAT signal_mappings.hdf5}

\item{pcr_signal}{Path to PCR signal_mappings.hdf5}

\item{out}{Output folder}

\item{reference_path}{Path to reference signal was mapped too}

\item{chunk_size}{Size of chunks (default 1e5)}

\item{max_cov}{Maximum coverage to include in motif detection (default: 200)}

\item{minPval}{Minimum p-value of current differences to considered event  (default: 1e-15)}

\item{hdbscan_args}{Arguments to HDBSCAN clustering (see documentation for possibilities)}

\item{umap_args}{Arguments to UMAP embedding (see documentation for possibilities)}

\item{align_event_sequences}{Align event sequence to increase possibility of detecting consensus sequence (default: TRUE)}

\item{iterations}{Number of clustering iterations (default 1)}

\item{min_entropy}{Minimum entropy of cluster alignment to consider a postion conserved (default: 1)}

\item{chunks_per_contig}{Number of chunks to process pr. contig (deafault: 0 (All chunks))}
}
\description{
Perform complete motif identification pipeline.
}
