% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste_path}
\alias{paste_path}
\title{Paste string to path}
\usage{
paste_path(...)
}
\arguments{
\item{...}{Strings to be concatenated}
}
\value{
Character
}
\description{
Ensures that leading or trailing / are correctly handled
}
