% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_chunk_current}
\alias{plot_chunk_current}
\title{Plot signal of chunk}
\usage{
plot_chunk_current(
  chunk,
  path_out,
  start_plot = "random",
  frame = 10,
  batch_size = 100
)
}
\arguments{
\item{chunk}{data.table, chunk with signal and pos_signal}

\item{path_out}{character, output path of plots}

\item{start_plot}{integer, start position of zoom plots (default: "random")}

\item{frame}{ineteger, Number of positions to include in zoom plots (default: 10)}

\item{batch_size}{integer, size of batching when estimating coverage (default: 100)}
}
\value{
nothing, save plot to path_out
}
\description{
Plots the signal and estimated coverage of a chunk.
}
