% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_detection.R
\name{plot_motifs}
\alias{plot_motifs}
\title{Visualise motif profiles}
\usage{
plot_motifs(
  cluster_path,
  plot_out,
  chunks_path,
  path_ref,
  n_extra_positions = 7,
  motifs_evaluated = "all"
)
}
\arguments{
\item{cluster_path}{path to motif_find clsuter.tsv}

\item{plot_out}{path to plot output}

\item{chunks_path}{path to preprocess chunks}

\item{n_extra_positions}{number of adjacent positions to include in plot}

\item{motifs_evaluated}{evaluation type (all, clustered, or vector of motifs)}
}
\value{
plots
}
\description{
Visualise the identified clusters
}
