% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prepare_metainfo}
\alias{prepare_metainfo}
\title{Run preprocess}
\usage{
prepare_metainfo(
  nat_mapping,
  pcr_mapping,
  nat_signal,
  pcr_signal,
  chunk_size = 1e+05,
  threads = 1,
  debug = FALSE,
  pvp = FALSE
)
}
\arguments{
\item{nat_mapping}{Path to NAT sorted bam file}

\item{pcr_mapping}{Path to PCR sorted bam file}

\item{nat_signal}{HDF5 object of NAT signal mappings}

\item{pcr_signal}{HDF5 object of PCR signal mappings}

\item{chunk_size}{Size to chunk reference into (default 1e4 bases)}

\item{threads}{Number of jobs to run parallel (default 1; >1 not supported on windows)}

\item{out}{Output path for preprocessed signal mappings}
}
\value{
metainfo list with nesated bacthes and types
}
\description{
Function to process signal and read mappings into preprocessed .tsv files, ready for current difference calculation
}
