% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Run preprocess}
\usage{
preprocess(
  nat_mapping,
  pcr_mapping,
  nat_signal,
  pcr_signal,
  out,
  chunk_size = 10000,
  threads = 1
)
}
\arguments{
\item{nat_mapping}{Path to NAT sorted bam file}

\item{pcr_mapping}{Path to PCR sorted bam file}

\item{nat_signal}{HDF5 object of NAT signal mappings}

\item{pcr_signal}{HDF5 object of PCR signal mappings}

\item{chunk_size}{Size to chunk reference into (default 1e4 bases)}

\item{threads}{Number of jobs to run parallel (default 1; >1 not supported on windows)}
}
\value{
preprocessed chunk saved to 'out'
}
\description{
Function to process signal and read mappings into preprocessed .tsv files, ready for current difference calculation
}
