% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_difference.R
\name{preprocess_all_chunks}
\alias{preprocess_all_chunks}
\title{Processes all chunks in metainfo list}
\usage{
preprocess_all_chunks(metainfo, h5_list, out, chunk_size = 1e+05, ...)
}
\arguments{
\item{h5_list}{list of nat and pcr h5 objects}

\item{out}{output folder of preprocessed chunks}

\item{chunk_size}{size of chunks}

\item{chunks_list}{list of batches with nat and pcr metainfo}

\item{plot_path}{path to output generated plot}
}
\value{
dt with calculated statistics
}
\description{
Loads signal and calculates current difference
}
