% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mappings.R
\name{preprocess_mappings}
\alias{preprocess_mappings}
\title{Loads and prepares read mappings for NAT and PCR}
\usage{
preprocess_mappings(nat, pcr, chunk_size)
}
\arguments{
\item{nat}{Path to NAT read mappings}

\item{pcr}{Path to PCR read mappings}

\item{chunk_size}{Size of chunks to divide contigs into}
}
\value{
data.table with read mappings
}
\description{
This function load metainformation of each read in a batch in a singal mapping hdf5 file.
}
