% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_difference.R
\name{process_chunk}
\alias{process_chunk}
\title{Process chunk}
\usage{
process_chunk(
  chunk_list,
  h5_list,
  chunk_size,
  plot_path = paste0("./current_plots/", unique(metainfo[[2]][[1]][[1]]$chunk_ref), "/")
)
}
\arguments{
\item{chunk_list}{list of batches with nat and pcr metainfo}

\item{h5_list}{list of nat and pcr h5 objects}

\item{chunk_size}{size of chunks}

\item{plot_path}{path to output generated plot}
}
\value{
dt with calculated statistics
}
\description{
Loads signal and calculates current difference
}
