% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{process_multichunk_reads}
\alias{process_multichunk_reads}
\title{Get read mappings that overextent their chunk}
\usage{
process_multichunk_reads(read_mapping, chunk_size)
}
\arguments{
\item{read_mapping}{Read mapping dt}

\item{chunk_size}{Size of chunks}
}
\value{
read mapping data.table
}
\description{
Gets the read that overextent their assigned chunk and return dt with new
chunk assignments
}
