% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mappings.R
\name{read_mapping_both}
\alias{read_mapping_both}
\title{Loads both NAT and PCR mappings}
\usage{
read_mapping_both(nat, pcr)
}
\arguments{
\item{nat}{Path to NAT read mappings}

\item{pcr}{Path to PCR read mappings}
}
\value{
read mapping data.table
}
\description{
Load both types of mappings at the same time and adds the corect type value. This is required for correct donsampling later.
}
