% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_metainfo_all}
\alias{read_metainfo_all}
\title{Read metainfo from signal_mappings.hdf}
\usage{
read_metainfo_all(hdf5)
}
\arguments{
\item{hdf5}{Open hdf5 object}
}
\value{
list of data.tables with metainformation of all batches
}
\description{
This function load all  metainformation in a singal mapping hdf5 file.
}
