% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replace_na_dt}
\alias{replace_na_dt}
\title{Replace all NA's in DT}
\usage{
replace_na_dt(DT, replacement = 0L)
}
\arguments{
\item{DT}{data.table}

\item{replacement}{value to replace with}
}
\value{
data.table with filled NA's
}
\description{
Replaces all the NA's in all columns in data.table with the specified value
}
