% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{unnest_dt}
\alias{unnest_dt}
\title{Unnests data.table}
\usage{
unnest_dt(dt, list_col, keep_col)
}
\arguments{
\item{dt}{data.table containing listed column}

\item{list_col}{Column that contain list}

\item{keep_col}{Columns that are kept in the returned data.table}
}
\value{
Unnested data.table
}
\description{
This function unnest list column of data.table.
}
