% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_detection.R
\name{visualise_clusters}
\alias{visualise_clusters}
\title{Visualise clusters}
\usage{
visualise_clusters(dt)
}
\arguments{
\item{dt}{data.table outputted from embed_and_cluster}
}
\value{
list(scatter_plot, logo_plot)
}
\description{
Visualise the identified clusters
}
