% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_index_to_vector}
\alias{add_index_to_vector}
\title{Return list with vector and indices}
\usage{
add_index_to_vector(increment_pos, vec)
}
\arguments{
\item{increment_pos}{Vector of positions of index increment}

\item{vec}{Vector to add indices too}
}
\value{
data.table with read mappings
}
\description{
Calculates index of one vector based on increment vector and returns list with both
}
