% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_signal_chunk}
\alias{add_signal_chunk}
\title{Adds signal to a complete chunk}
\usage{
add_signal_chunk(metainfo_chunk, hdf5)
}
\arguments{
\item{metainfo_chunk}{a list of batches, with NAt and PCR metainfo}

\item{hdf5}{a list of nat and pcr hdf5 objects}
}
\value{
logical of success status (signal is loaded in memory)
}
\description{
Iterator for adding signal to all bacthes and type (nat and pcr) of a
chunk of the chunk list outputted by prepare_metainfo
}
