% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_difference.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Calculates statistics between NAT and PCR signal}
\usage{
calculate_statistics(signal, window = 8, value_col = "mean_diff")
}
\arguments{
\item{signal}{Preprocessed signal mappings}

\item{window}{Number of position to in frame include at each side of event}

\item{value_col}{Statistic to include in event frame}
}
\value{
current difference data.table
}
\description{
Calculates statistics between NAT and PCR signal
}
