% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mappings.R
\name{downsample}
\alias{downsample}
\title{Downsample either nat or pcr to least abundant type}
\usage{
downsample(read_mapping, chunk_size, min_cov = 20)
}
\arguments{
\item{read_mapping}{Preprocessed signal mappings}

\item{chunk_size}{size of chunks}

\item{min_cov}{Minimum coverage of type}
}
\value{
read mapping data.table
}
\description{
Removes read of the most bundant mapping type, such there are almost equal mapped bases of each type to each chunk.
}
