% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_difference.R
\name{get_event_features}
\alias{get_event_features}
\title{Extract features from event statistics}
\usage{
get_event_features(signal, p_value_threshold = 1e-10)
}
\arguments{
\item{signal}{Preprocessed signal mappings}

\item{p_value_threshold}{Threshold for event selection (based on two sided wilcox test)}
}
\value{
data.table with feature columns and reference position
}
\description{
Selects events based on p-value and extract feature from signal dt
with calculated statistics, which is then ready for embedding and clustering
}
