% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_reference_context}
\alias{get_reference_context}
\title{Load signal mappings of one batch}
\usage{
get_reference_context(signal, mapping)
}
\arguments{
\item{signal}{data.table of metainfo loaded with read_metainfo}

\item{mapping}{Read mappings dt}
}
\value{
Nothing, signal mappings will be added to the metainfo object in memory
}
\description{
Loads the signal mappings associated with a batch
}
