% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prepare_metainfo}
\alias{prepare_metainfo}
\title{Prepare chunks for signal loading}
\usage{
prepare_metainfo(
  nat_mapping,
  pcr_mapping,
  nat_hdf5,
  pcr_hdf5,
  chunk_size = 1e+05,
  threads = 1,
  debug = FALSE,
  pvp = FALSE
)
}
\arguments{
\item{nat_mapping}{Path to NAT sorted bam file}

\item{pcr_mapping}{Path to PCR sorted bam file}

\item{chunk_size}{Size to chunk reference into (default 1e5)}

\item{debug}{logical, print additional information}

\item{pvp}{logical, load only PCR and split it into 1:1 labeled PCR and NAT}

\item{hdf5}{list of nat and pcr hdf5 objects}
}
\value{
metainfo list with nested bacthes and types
}
\description{
Function to process signal and read mappings,
ready for signal loading
}
